Codecs refactoring. Move common codecs to their own file. Harmonize use of val vs def for fields.
